Cut analytics days in Riyadh time, and count in numbers
Days were bucketed on UTC, so a Riyadh designer's day ended at 3am and
every evening's activity — when portfolio links actually get shared —
landed on the next day's row.

Both engines that cut a day now agree: Intl in Node for the keys we write,
to_char(... AT TIME ZONE ...) in Postgres for the ones we read back, with a
test pinning them together across the midnight boundary.

Postgres also returns COUNT and SUM as bigint, which pg hands back as a
string, so the series values were strings and the charts were adding up
text. Cast in SQL rather than in the coercion table: a `value` suffix there
would also have caught settings.value, which is JSON and must stay a string.

+++ src/lib/invitations.ts
@@ -132,10 +132,10 @@ export async function revokeInvitation(id: string) {
 
 export async function invitationStats() {
   const row = await get<{ total: number; redeemed: number; live: number }>(
-    `SELECT COUNT(*) AS total,
-            COALESCE(SUM(used_count), 0) AS redeemed,
+    `SELECT COUNT(*)::int AS total,
+            COALESCE(SUM(used_count), 0)::int AS redeemed,
             SUM(CASE WHEN revoked = 0 AND used_count < max_uses
-                      AND (expires_at IS NULL OR expires_at > ?) THEN 1 ELSE 0 END) AS live
+                      AND (expires_at IS NULL OR expires_at > ?) THEN 1 ELSE 0 END)::int AS live
        FROM invitations`,
     now(),
   );
